requirements.txt
dnspython==1.12.0
ipwhois==0.10.2
ipaddr==2.1.11
pythonwhois==2.4.3
futures == 3.0.3
xmlutils == 1.1
lxml == 3.7.2
